+++ docs/source/NEWS.rst
@@ -19,6 +19,7 @@ Features
   - Limitations:
     - No session options are supported for `start_session()` yet
     - Only `write_concern` and `max_commit_time_ms` options are supported for `start_transaction()`
+- Minimum Twisted version is now 22.8.0
 
 API Changes
 ^^^^^^^^^^^